Replace ESP install component (#618)
* Remove the ESP install component code

* Add esp-web-tools NPM package

* Remove esp-web-tools NPM package

* Add js script to Docusaurus

* Refactor component code

* Small refactor on global declare

* Rename the file and component
