+++ docs/docs/getting-started.mdx
@@ -6,7 +6,7 @@ description: 'Instruction to get started with the Home Assistant Glow'
 
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
-import EspInstallButton from '@site/src/components/EspInstallButton';
+import ESPHomeButton from '@site/src/components/ESPWebTools';
 import { faUsb } from '@fortawesome/free-brands-svg-icons';
 
 # Get started
@@ -63,7 +63,7 @@ If your device does not appear in the list, make sure you are using a USB data c
       - The Home Assistant Glow is now connected to your network.
 6. Congratulations 🎉 You have successfully installed the firmware on your Home Assistant Glow.
 
-<EspInstallButton unsupportedMessage="Provisioning via USB only works in browsers which support Web Serial, such as Chrome or Edge." />
+  <ESPHomeButton />
   </TabItem>
 </Tabs>
